requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
kiutils>=1.4.0
cairosvg>=2.7.0
pillow>=10.0.0
